moved matrix test
